backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
groq==0.4.2
python-dateutil==2.8.2 